Remove bullshit torrent clients
Why are other torrent clients removed?

There is a simple explanation, they don't have the ability to bind a specific IP/adapter or network interface which is the only reliable solution to prevent possible leaks in case you VPN fails to bind an IP address. This ensures 100% that only the VPN connection is allowed to download something.