Insecure websites removed (soft banned)
What is a soft ban?
- The removed websites with an insecure protocol are now listed under banned.md, however if the webmaster/owner decides to upgrade the website to a secure and better standard then I'm going to re-list them on the Overview page.

Why are some pages (which are insecure) marked/tagged with `insecure`?
- Some very popular websites are insecure, however they are not been removed for several reasons e.g. popularity, search reasons etc.

Website with "beta" https://?
Some websites can be reached via HTTPS-Everywhere & Co. to "upgrade" (redirect) the website to https://, however if the website does not naively support HTTPS by default then it's considerable a "beta" redirection which is not been tolerated. No one should be forced to use extensions/scripts/filters to reach a secure website. Welcome to 2019.

Minimum website standard?
TLS 2.0+. According to several stats pages only <5% of all encrypted pages using an outdated TLS version 1.0/1.1/1.2 etc. So welcome in Russia my friends... Just update your standard. It's not hard, if you can't do it then why should we trust your page/content?!

Why does it matter if the website only lists/links to other secure providers e.g. zippyshare (sharehoster) etc.?!
The reason is that Warez pages are been actively monitored and indexed, the risk is high that those pages are somehow compromised (XSS, MITM, etc) in order to find out who is going to click on link x or simply to track you down.

Wrong assumption (?)
I do not support or vouch for any https site here, but in my point of view in 2019 every freak'n page should be at least be in TLS 2.0 (or higher). I doubt every website owner/site or hoster which is not capable of providing a secure website/content, especially if it comes to Warez.