Link to other websites anonymously
We ensure that (in case you come from another website like GitHub, CornHub & Co) the links you clicked here are "obfuscated" (I know it's the wrong word for it). This basically prevents third-parties and first-parties to back-track your site you came from. Theoretically this also answers the liability questions for the links here on this repository because the provider is responsible and if we "give" another provider the real link then he is responsible for the content. However, I'm not sure if that is the case here because github explicitly looks for https://. 

There are also browser extensions to do that and some Browsers like Brave or Firefox (with some config tweaks) can do this without any extensions, but since I don't want to force anyone to install an extension or to enable blocking the referrer by default.